Fix bug in rocket page
