+++ src/App.js
@@ -2,8 +2,18 @@ import { Route, Routes } from 'react-router-dom';
 import MainHeader from './components/MainHeader';
 import MissionsPage from './pages/MissionsPage';
 import RocketsPage from './pages/RocketsPage';
+import { useEffect } from 'react';
+import { useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
+import initState from '../redux/rockets/rocketsAPI';
 
 function App() {
+  const dispatch = useDispatch();
+  const state = useSelector((state) => state.rockets.rocketsData);
+
+  useEffect(() => {
+    dispatch(initState());
+  }, [dispatch]);
   return (
     <>
       <header>
@@ -12,7 +22,7 @@ function App() {
 
       <main className="container my-3">
         <Routes>
-          <Route path="/" element={<RocketsPage />} />
+          <Route path="/" element={<RocketsPage state={state} />} />
           <Route path="/missions" element={<MissionsPage />} />
         </Routes>
       </main>
