+++ src/pages/RocketsPage.jsx
@@ -1,17 +1,6 @@
-import { useEffect } from 'react';
-import { useSelector } from 'react-redux';
-import { useDispatch } from 'react-redux';
 import Rockets from '../components/rockets/Rockets';
-import initState from '../redux/rockets/rocketsAPI';
 
-const RocketsPage = () => {
-  const dispatch = useDispatch();
-  const state = useSelector((state) => state.rockets.rocketsData);
-
-  useEffect(() => {
-    dispatch(initState());
-  }, [dispatch]);
-
+const RocketsPage = ({ state }) => {
   return <Rockets list={state} />;
 };
